+++ tests/testthat/test-translate-sql.R
@@ -153,6 +153,10 @@ test_that("different arguments of substr are corrected", {
   expect_equal(translate_sql(substr(x, 3, 1)), sql("SUBSTR(`x`, 3, 0)"))
 })
 
+test_that("substring is also translated", {
+  expect_equal(translate_sql(substring(x, 3, 4)), sql("SUBSTR(`x`, 3, 2)"))
+})
+
 test_that("paste() translated to CONCAT_WS", {
   expect_equal(translate_sql(paste0(x, y)),             sql("CONCAT_WS('', `x`, `y`)"))
   expect_equal(translate_sql(paste(x, y)),              sql("CONCAT_WS(' ', `x`, `y`)"))